read and wishlist button function finished

+++ src/utiles/index.js
@@ -1,4 +1,4 @@
-import toast from "react-hot-toast"
+
 export const getBooks = () => {
     let books = []
     const storedBooks = localStorage.getItem('books')
@@ -8,14 +8,20 @@ export const getBooks = () => {
     return books
 }
 
-export const saveBooks = bookId => {
-    let books = getBooks()
-    const isExist = books.find(b => b === bookId)
-    if (!isExist) {
-        books.push(bookId)
-        localStorage.setItem('books', JSON.stringify(books))
-        return toast.success('Book read Successfully!')
-    }
-    toast.error('Already Read')
+export const saveBooks = (bookId, action) => {
+    let books = getBooks();
+    const isExist = books.find(b => b === bookId);
     
-} 
+    if (action === 'read' && !isExist) {
+        books.push(bookId);
+        localStorage.setItem('books', JSON.stringify(books));
+        return true; // Book marked as read successfully
+    } else if (action === 'wishlist' && !isExist) {
+        books.push(bookId);
+        localStorage.setItem('books', JSON.stringify(books));
+        return true; // Book added to wishlist successfully
+    } else {
+        return false; // Book already marked as read or added to wishlist
+    }
+};
+
